Set document.innerHTML to <video src='mp4videoSrc' /> upon resolution
For some reason, navigating to the videoSrc URL from mp4upload triggers the browser to download the video. To prevent this, and play the video in the browser instead, set document.innerHTML to a video manually. Note: the 'referrer' must be removed for this to work due that mp4upload's videos are supposed to be nested within an iframe that points to an mp4upload URL