Move fetch/CORS informational comment to 'fetchCors()' JSDoc
This improves the bookmarking process because it allows simply selecting the entire file and adding it as a bookmark instead of having to manually select everything except the comment at the end. Putting it in the fetchCors() JSDoc also adds context around what was attempted and where to pick up if wanting to continue attempts to deal with CORS issues. Finally, shorten the 'Cookie' header and add a comment for describing the purpose of the headers block.