Rename bash-subsystem txt file to a .profile file and fix toWindowsPath()
Renaming the file allows it to be 'source'd in the bash subsystem and also clarifies its purpose. Fix the toWindowsPath() function to (1) replace '/mnt/c' with 'C:' for when the shell is in a Windows directory, and (2) prepend '$rootdir/' to the path if the shell is in a linux subsystem directory. Finally, merge 'export rootdir' line with rootdir declaration and add homedir declaration for use in Windows programs.