Update kissanime > novelplanet scraper to work
Novelplanet requires the use of cookies for their '/api/source' API to work. Since it's loaded in an iframe, the cookies aren't accessible to the parent document, and since the needed cookies are marked as 'secure', they can't be read in JavaScript. Thus, convert the Novelplanet function to redirect from kissanime to novelplanet, then fetch the video URL and redirect from novelplanet to the video URL. Also, update function name, comments, and 'async' function signature