Change Sublime JS-beautify brace-style to preserve braces intentionally put inline
Helps for minor formatting of oddly-structured objects. I don't remember exactly what was causing the issue, but both `end-expand` and `collapse-preserve-inline` result in arrays of objects having the first opening object brace on the same line as the opening array bracket, and `}, {` array entries remaining on the same line; so might as well keep the default behavior.