Update fstab entry to allow execution of NTFS files
When trying to run any nested executable on NTFS drives, it fails because the nested (spawned) child process is not able to execute certain files. A key example is webpack, where node/npm tries to run webpack-dev-server, but to do so it requires spawning a child process which fails because it doesn't have execution rights. Thus, to be able to execute NTFS files in child processes, change the permissions from `-rw-rw-rw` to `rwx-rw-rw`. Note how this allows the user to execute files but not the group or public.