Convert egrep to 'grep -E' to allow for standard regex input
egrep, i.e. 'grep -e', requires all special characters to be escaped even if they're part of the regex expression, e.g. '(a|\(b\))' has to be '\(a\|\(b\)\)'. This is annoying as it doesn't distinguish between the regex's special characters and actually-escaped characters. Thus, convert egrep to 'grep -E' where the -E flag converts from 'regex' to 'extended regex'