Fix Ctrl+Click issues in JetBrains IDEs
The real problem here is Mac doesn't allow you to disable their built-in mapping where Ctrl+Left-Click opens the context (right-click) menu. Even though Ctrl+Click was remapped to Command+Click everywhere except JetBrains and JetBrains key combos are (for the most part) working as intended, the Ctrl+Click functionality was working in JetBrains, but it'd be followed up with a right click after the "Go to Declarations & Usages" event was triggered within the IDE.

Fix this by remapping Ctrl+Click --> Command+Click in Karabiner and changing the key combo in JetBrains in the same way.