Remove unused files
Since there are no simple key remap or other configs in Karabiner right now, remove the karabiner.json (which is the only file that's actually read by Karabiner, the rest are just for importing to/exporting from that file) file. Also, remove the backup files since they're just noise (not to mention that Git is a backup system in and of itself).