Allow Command+Click to be used separately from Ctrl+Click
By setting a variable for tracking which modifier key was pressed in the key combo, we can now distinguish between Ctrl+Click and Command+Click in programs like JetBrains where Ctrl+Click was mapped to Command+Click. This is highly useful because now Ctrl+Click and Command+Click can be mapped to separate functions in JetBrains, whereas before, mapping Command+Click to Option+Click would cause both Ctrl+Click (which is forwarded to Command+Click) and Command+Click to be mapped to Option+Click.

Similarly, add the aforementioned Command+Click --> Option+Click mapping so it can be used in JetBrains.

Note: The `to.set_variable` entry **must** be declared before the `to.key_code` entry because otherwise, the key_code will be pressed and released immediately (due to Karabiner triggering the `to` entries sequentially). Putting set_variable above key_code fixes this by allowing the key_code to be the last action, so holding the key keeps it held down as expected.