Only source linux's profile if the dir exists
Since my Linux system has the Documents/ dir on a separate partition, it may not be mounted on boot, causing the OS to throw errors when trying to boot. Since the partition will be mounted by the time login is finished and all new shells opened up after that time will source the profile as expected, simply add a check for this boot edge case.