Finally a way to return arrays in bash! Update array.slice() to do so if flag is passed
If the flag is present, then "return" the array; otherwise, echo it to stdout (useful for if this function is called from the live shell).

Actually, the array isn't technically returned, rather the calling parent function must pass in a return-array name (doesn't have to be declared first), and then this function can populate it. Note: This logic has to be done in this function, otherwise any quoted string entries will be unquoted when returned (e.g. `echo "${myArray[@]}").

TODO Update `array.values-from-name()` to do this as well.