Source bash profiles/utils only if they're found
When the `dotfiles` repo is on an external partition, Linux sometimes has a hard time finding it. More specifically, the aliases don't seem to work even though **they're defined above any `source` calls**. Not sure if this is a bug or not, but it can be solved by simply checking if the alias `thisDir` is defined and, if so, only then proceed with `source`ing all the other files.