Make 'ls --color' the default for all OSes
Now that mac is using Bash >= v4, it both supports and needs the '--color' flag to display 'ls' content with color. Thus, default 'ls' aliases to include the '--color' flag since it's needed in all environments (and the only reason not to use it was b/c Mac is stupid and uses Bash < 4 by defulat).