Update WebStorm to inherit user's PATH for custom Node locations
For some reason, the newest update broke WebStorm's ability to find the Node binary. Fix that by making the program open from a bash login shell (-l) with the fallback of making it interactive (-i). Also, update the Terminal settings within WebStorm itself so that the terminal it uses in the 'Terminal' tab (not in run configurations) is also a login shell rather than a plain old /bin/bash.

See:

* [Solution](https://youtrack.jetbrains.com/issue/IDEABKL-7589)
    - [Previous issue linking to it](https://intellij-support.jetbrains.com/hc/en-us/community/posts/205437150-Terminal-not-sourcing-bash-profile?page=1#community_comment_360000167679)
* [Original issue](https://intellij-support.jetbrains.com/hc/en-us/community/posts/360008275400-Node-Interpreter-node-No-executable-found-in-PATH-)
* Others experiencing the same thing:
    - [1](https://intellij-support.jetbrains.com/hc/en-us/community/posts/208567485-Pycharm-terminal-is-missing-part-of-PATH)
    - [2](https://youtrack.jetbrains.com/issue/PY-17816)
* [Make inner-IDE terminal tabs use login shell](https://stackoverflow.com/questions/36592226/bashrc-not-sourced-on-intellij-ideas-terminal)