Make Sublime Text open with login shell to preserve PATH
Like with WebStorm, this allows the correct Java, JavaScript, etc. languages to be used in Sublime. Note: Opening Sublime in a logged-in terminal via 'subl' did actually preserve PATH, it was solely the times when Sublime was opened via double-clicking on a file (.desktop), opened from the super/start menu (.desktop), from right-click menu (.nemo_action), or keyboard shortcut (dconf) that it wouldn't use the custom .profile PATH. Thus, fix it in all 3 places so it's ubiquitous.