Add new 'Java' Sublime build
This way, on new systems, I won't have to copy the 'Packages/Java' directory in order to use the better Java build system (though it's still recommended to copy it so JavaC.sublime-build will still run the file). Using the name 'Java' distinguishes it from JavaC.sublime-build (the built-in one) so it'll be clear which is which. Also, add additional 'JavaDoc' and 'JAR' build systems but nest them inside the 'Java' one (displayed as 'Java - JavaDoc').

Refs:

* [Build Systems (v4)](https://www.sublimetext.com/docs/build_systems.html)
* [Build Systems (v2)](https://sublime-text-unofficial-documentation.readthedocs.io/en/sublime-text-2/reference/build_systems.html)
    - While old, they still have good information forgotten in the v4 docs