Minor: Change parseArgs USAGE string to use special Bash var
Since this function is going to be the basis for all option, arg, and USAGE handling, it should be very descriptive in its source code. As such, use this special Bash variable to show how new functions calling this can have the command name in USAGE automatically updated in the event of renaming it.