Change Sublime go-to-any-definition key binding to match the others
All the other definition/reference-based key bindings use ALT instead of CTRL; the only one using CTRL is go-to-file. Thus, change the go-to-any-definition key binding to be similar to the others: Use ALT for definitions and Shift for enhancing the definition search. Also, update comment to better reflect that this means go-to-file instead of go-to-anything-including-files.