fe-start: Don't read from stdin; Default to only docker
'-' is a special character in bash to tell the command to read from stdin, so replace it with '.' instead. Also, default the function to only start the Docker containers since `yarn build` is often run from the IDE; However, if '.' is passed, then remove it and start `yarn build --watch` to feign not passing any args at all.