Group `ls` alias output by dirs first, then by files
After trying many different permutations, it seems we can either (1) have dotfiles at the top of the list, or (2) have directories at the top of the list. For (1), the output seems to be case-insensitive, but for (2) they are case-sensitive. TODO Both of these (dotfiles at the top of their respective file/dir group and output being case-sensitive) need to be fixed to allow both.