Add new git submodule for my personal content
In preparation for making my dotfiles themselves open source so they can improve the lives of others, I first need to make it easy for me to be able to use my own personal content while preventing others from being able to see them.

Git submodules can be selectively cloned or not. As such, this means that I can clone it to have access to my personal content but others can't.