Minor bug fixes: Always declare `local OPTIND`; Simplify `parseArgs` return-val checks
In some cases, the value of OPTIND was carrying forward between functions; fix that by re-declaring it in every function that uses it.

Likewise, do a bit of minor cleanup by declaring OPTIND right before the `getopts` usage so that the variables more important/related to the function are declared first to improve readability.