Bug fix: Don't delete the first unknown var in `parseArgs()` if breaking out of it
In the event that the parent function calling `parseArgs` option flags are forwarded to an underlying script with its own options, we don't want `parseArgs()` to inhibit the ability to keep those flags. Thus, if `break` is passed as an unknown-flag handler, move OPTIND backwards to keep that unknown flag in the returned `argsArray`.