Only print error in root .profile if in live shell
This allows the error to be printed if the OS-specific directory isn't specified while still hiding the error from custom scripts, shebangs, etc. Using this logic, we only show the error if the user called `source ${dotfilesDir}/.profile` without the nested platform directory either in their shell manually or in their `$HOME/.profile`. That way, we can still warn them that they need to specify a directory without interfering with custom scripts' usage of this repo.