Harden the `isBeingSourced()` logic to be more robust
Now that we know a reliable way to get the file in question, as opposed to assuming which index in the `BASH_SOURCE` array the file resides, use that first and fallback to `BASH_SOURCE` only if that fails. Also, update the USAGE string accordingly and make the code more readable by extracting some logic to new variables.