Minor: Remove superfluous if-statement
Now that `shopt -s expand_aliases` is set in the root `.profile`, the `if type thisDir &>/dev/null` check shouldn't be necessary.

That if-statement was introduced in commit 3b24cd6c3920012286222716c268e4acc119391c ("Source bash profiles/utils only if they're found") where it was likely the issue was because aliases weren't being exported, not that they weren't defined; this is likely safe to assume based on the logic that the `thisDir` alias is defined literally right above the if-statement. Thus, remove it to simplify the `.profile` code.