Add new JS util to get chapter titles from YouTube
YouTube is stupid and doesn't give a list of chapters, making it very difficult for users to know where the specific chapter they want exists in videos with many chapters. This function makes it much easier to find the index the desired chapter exists such that users can quickly navigate to it.