Bug fix: Fix arg-passing logic for git-ignore-paths function
The problem ended up being (1) that `':(top)' ':(exclude)my-file'` was being used instead of `':(top,exclude)my-file'` and (2) quoting `"':(exclude)file-a' ':(exclude)file-b'"` (due to parent functions calling this function to use `<cmd> "$(gitIgnoreDirsArgs "$@")"`). Fix both of them in one fell swoop in this commit, adding a help-doc entry to ensure it doesn't happen with new functions in the future.