Add function to ensure a complete installation of `store-checkout-components`
Not sure if it's the WSL-Windows conflicts, Maven configurations causing grandchild-projects to not inherit grandparent-${project.version}, or just the fact that I'm trying to use WSL instead of only running on Windows (unlikely since I've mapped the WSL `mvn` command to the Windows installation of Maven), but no matter what I do, I get dependency-version conflicts which don't correctly interpret `<version>(1.2.3,${project.version}]</version>`.

I've debugged this a hundred times over, but the simplest (and likely *only*) way to fix it is to change the minimum value of that range to the same minimum value of the version of the grandparent.

Thus, add a new function to use `sed` to extract the min version number from the grandparent and then apply it to all (grand)children projects within the repository. This will make `mvn [clean] install` actually work rather than throwing hundreds of errors for no reason.