Make PostgreSQL DB creation logic default to whomever owns the PGDATA dir
`PGDATA=/var/lib/pgsql/data` doesn't allow us to call `initdb` for some reason (and `sudo -u postgres` doesn't fix it) so for ease of use, we make the PGDATA owned by the current user. We make the directory with respective access permissions.

Then, we update `initdb` and `postgresStart` to use the correct root user.

Finally, we default some variables, like `_pgdataDir` to `$PGDATA` instead of throwing an error.